% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510package.R
\name{adj_R2}
\alias{adj_R2}
\title{Calculate the adjusted R-squared statistic between predicted and ground truth continuous variables}
\usage{
adj_R2(pred, truth, n_p)
}
\arguments{
\item{pred}{A numeric vector of predicted values}

\item{truth}{A numeric vector of true values}

\item{n_p}{Number of model parameters, excluding the intercept}
}
\value{
Adjusted R-squared statistic
}
\description{
Calculate the adjusted R-squared statistic between predicted and ground truth continuous variables
}
\examples{
pred <- c(1, 2, 3, 4, 5)
truth <- c(1.2, 2.1, 3.3, 4.4, 4.8)
adj_R2(pred, truth, 2)
pred <- rnorm(100)
truth <- rnorm(100, mean = 3)
adj_R2(pred, truth, 4)
pred <- runif(50)
truth <- runif(50)
adj_R2(pred, truth, 1)
}
