% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510package.R
\name{classes}
\alias{classes}
\title{Return the classes of each variable in a tibble}
\usage{
classes(x)
}
\arguments{
\item{x}{A tibble for which to return the classes of each variable}
}
\value{
A character vector containing the classes of each variable in a tibble x
}
\description{
Return the classes of each variable in a tibble
}
\examples{
classes(iris)
classes(mtcars)
classes(data.frame(x = 1:10, y = letters[1:10], z = factor(1:10)))
}
