% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510package.R
\name{df_scale}
\alias{df_scale}
\title{Scale numeric variables in a tibble}
\usage{
df_scale(x, center = T, scale = T)
}
\arguments{
\item{x}{A tibble to scale}

\item{center}{Logical, whether to center the numeric variables (default is TRUE)}

\item{scale}{Logical, whether to scale the numeric variables (default is TRUE)}
}
\value{
A tibble x in which the numeric variables have been scaled
}
\description{
Scale numeric variables in a tibble
}
\examples{
df_scale(iris)
df_scale(mtcars, center = F)
df_scale(data.frame(x = rnorm(10), y = rnorm(10, mean = 5), z = rnorm(10, sd = 2)), scale = F)
}
