% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510package.R
\name{is_max}
\alias{is_max}
\title{Check if an element in a vector is equal to the maximum value}
\usage{
is_max(x, na.rm = T)
}
\arguments{
\item{x}{An atomic vector to check for maximum values}

\item{na.rm}{Logical, whether to remove NA values before checking (default is TRUE)}
}
\value{
A logical vector with TRUE where x equals its maximum value
}
\description{
Check if an element in a vector is equal to the maximum value
}
\examples{
is_max(c(1, 2, 3, 1))
is_max(c(5, NA, 2, 7, 2), na.rm = F)
is_max(rnorm(10))
}
