% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510package.R
\name{log_likelihood_chisq}
\alias{log_likelihood_chisq}
\title{Calculate the log-likelihood of a sample under the chi-squared density}
\usage{
log_likelihood_chisq(x, df)
}
\arguments{
\item{x}{A numeric vector of observations}

\item{df}{Degrees of freedom of the chi-squared distribution}
}
\value{
The log-likelihood of the sample under the chi-squared density
}
\description{
Calculate the log-likelihood of a sample under the chi-squared density
}
\examples{
log_likelihood_chisq(rchisq(100, df = 3), 3)
log_likelihood_chisq(rchisq(100, df = 10), 10)
log_likelihood_chisq(c(1, 2, 3, 4, 5), 2)
}
