% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510package.R
\name{log_likelihood_f}
\alias{log_likelihood_f}
\title{Calculate the log-likelihood of a sample under the F density}
\usage{
log_likelihood_f(x, df1, df2)
}
\arguments{
\item{x}{A numeric vector of observations}

\item{df1}{Degrees of freedom for the numerator of the F distribution}

\item{df2}{Degrees of freedom for the denominator of the F distribution}
}
\value{
The log-likelihood of the sample under the F density
}
\description{
Calculate the log-likelihood of a sample under the F density
}
\examples{
log_likelihood_f(rf(100, df1 = 3, df2 = 5), 3, 5)
log_likelihood_f(rf(100, df1 = 10, df2 = 20), 10, 20)
log_likelihood_f(c(1, 2, 3, 4, 5), 2, 4)
}
