% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510package.R
\name{log_likelihood_norm}
\alias{log_likelihood_norm}
\title{Calculate the log-likelihood of a sample under the normal density}
\usage{
log_likelihood_norm(x, mean, sd)
}
\arguments{
\item{x}{A numeric vector of observations}

\item{mean}{Mean of the normal distribution}

\item{sd}{Standard deviation of the normal distribution}
}
\value{
The log-likelihood of the sample under the normal density
}
\description{
Calculate the log-likelihood of a sample under the normal density
}
\examples{
log_likelihood_norm(rnorm(100), 0, 1)
log_likelihood_norm(rnorm(100, mean = 5, sd = 2), 5, 2)
log_likelihood_norm(c(1, 2, 3, 4, 5), 3, 1)
}
