% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510package.R
\name{log_likelihood_t}
\alias{log_likelihood_t}
\title{Calculate the log-likelihood of a sample under the t density}
\usage{
log_likelihood_t(x, df)
}
\arguments{
\item{x}{A numeric vector of observations}

\item{df}{Degrees of freedom of the t distribution}
}
\value{
The log-likelihood of the sample under the t density
}
\description{
Calculate the log-likelihood of a sample under the t density
}
\examples{
log_likelihood_t(rt(100, df = 3), 3)
log_likelihood_t(rt(100, df = 10), 10)
log_likelihood_t(c(1, 2, 3, 4, 5), 2)
}
