% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510package.R
\name{r2}
\alias{r2}
\title{Calculate the R-squared statistic between predicted and ground truth continuous variables}
\usage{
r2(pred, truth)
}
\arguments{
\item{pred}{A numeric vector of predicted values}

\item{truth}{A numeric vector of true values}
}
\value{
R-squared statistic
}
\description{
Calculate the R-squared statistic between predicted and ground truth continuous variables
}
\examples{
pred <- c(1, 2, 3, 4, 5)
truth <- c(1.2, 2.1, 3.3, 4.4, 4.8)
r2(pred, truth)
pred <- rnorm(100)
truth <- rnorm(100, mean = 3)
r2(pred, truth)
pred <- runif(50)
truth <- runif(50)
r2(pred, truth)
}
