% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510package.R
\name{rando}
\alias{rando}
\title{Randomly sample elements or rows from a vector or dataframe-like object}
\usage{
rando(x, n = 1, replace = T)
}
\arguments{
\item{x}{An atomic vector or dataframe-like object to sample from}

\item{n}{Number of samples to return (default is 1)}

\item{replace}{Logical, whether to sample with replacement (default is TRUE)}
}
\value{
A vector or dataframe with n samples or rows, depending on the input type
}
\description{
Randomly sample elements or rows from a vector or dataframe-like object
}
\examples{
rando(1:10, 5)
rando(mtcars, 3)
rando(letters, 3, replace = F)
}
