% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510package.R
\name{rep_mat}
\alias{rep_mat}
\title{Replicate the rows and columns of a matrix or dataframe}
\usage{
rep_mat(x, M = 1, N = 1)
}
\arguments{
\item{x}{A matrix or dataframe to replicate}

\item{M}{Number of times to replicate the rows (default is 1)}

\item{N}{Number of times to replicate the columns (default is 1)}
}
\value{
A matrix created by replicating the rows and columns M and N times
}
\description{
Replicate the rows and columns of a matrix or dataframe
}
\examples{
rep_mat(matrix(1:4, nrow = 2), M = 2, N = 2)
rep_mat(mtcars[1:5, 1:3], M = 3)
rep_mat(matrix(rnorm(9), nrow = 3), N = 4)
}
