% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510package.R
\name{specificity}
\alias{specificity}
\title{Calculate the specificity of a binary classifier}
\usage{
specificity(pred, truth)
}
\arguments{
\item{pred}{A binary vector of predicted labels}

\item{truth}{A binary vector of true labels}
}
\value{
Specificity of the classifier
}
\description{
Calculate the specificity of a binary classifier
}
\examples{
pred <- factor(c(1, 0, 1, 1, 0))
truth <- factor(c(1, 0, 1, 0, 0))
specificity(pred, truth)
pred <- rbinom(100, 1, 0.5)
truth <- rbinom(100, 1, 0.7)
specificity(pred, truth)
pred <- factor(c(1, 0, 1, 1, 0, 1, 0, 1))
truth <- factor(c(1, 1, 1, 1, 0,0, 0, 0))
specificity(pred, truth)
}
